Made some restructurisation in the project
