+++ backend/internal/transport/input.go
@@ -1,0 +1,17 @@
+package transport
+
+type SignUpInput struct {
+	Username string `json:"name" validate:"required"`
+	Email    string `json:"email" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
+
+type SignInInputUsername struct {
+	Username string `json:"username" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
+
+type SignInInputEmail struct {
+	Email    string `json:"email" validate:"required"`
+	Password string `json:"password" validate:"required"`
+} // чтобы потом у юзверя был выбор - логиниться через почту или юз
